+++ google2/CHANGELOG.md
@@ -13,7 +13,8 @@ Changed:
 - Set `goog-compute-%d` threads as daemon so that they won't prevent JVM from shutdown
 - Set `goog-publisher-%d` and `goog-subscriber-%d` threads as daemon so that they won't prevent JVM from shutdown
 - Set `goog-dataproc-%d` and Set `goog-disk-%d` as daemon threads for dataproc and disk client so that threads used by these 2 clients have more readable names
- 
+- Change return type for `GoogleComputeService`'s `modifyInstanceMetadata`, `addInstanceMetadata` and `removeInstanceMetadata` to `Option[Operation]`
+
 Dependency Upgrades:
 | Dependency   |      Old Version      |  New Version |
 |----------|:-------------:|------:|
@@ -35,7 +36,7 @@ Dependency Upgrades:
 | client-java |   12.0.0   |   14.0.0 |
 | cats-effect |   3.3.1   |   3.3.2 |
 
-SBT Dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.23-95fdde6"`
+SBT Dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.23-TRAVIS-REPLACE-ME"`
 
 ## 0.22
 Breaking Changes: